feat: move initialize, history, utils, config, and pubSub to top level exports
