--- src/config/index.js
@@ -1,12 +1,0 @@
-export {
-  configure,
-  getConfig,
-  setConfig,
-  mergeConfig,
-  ensureConfig,
-  getConfigService,
-  resetConfigService,
-  CONFIG_TOPIC,
-  CONFIG_CHANGED,
-} from './interface';
-export { default as ProcessEnvConfigService } from './ProcessEnvConfigService';
