--- src/init/index.js
@@ -1,15 +1,0 @@
-export {
-  APP_TOPIC,
-  APP_PUBSUB_INITIALIZED,
-  APP_CONFIG_INITIALIZED,
-  APP_AUTH_INITIALIZED,
-  APP_I18N_INITIALIZED,
-  APP_LOGGING_INITIALIZED,
-  APP_ANALYTICS_INITIALIZED,
-  APP_READY,
-  APP_INIT_ERROR,
-  history,
-  initError,
-  auth,
-  initialize,
-} from './interface';
