--- src/pubSub/PubSubJsService.js
@@ -1,16 +1,0 @@
-
-import PubSub from 'pubsub-js';
-
-export default class PubSubJsService {
-  subscribe(type, callback) {
-    return PubSub.subscribe(type, callback);
-  }
-
-  unsubscribe(token) {
-    PubSub.unsubscribe(token);
-  }
-
-  publish(type, data) {
-    PubSub.publish(type, data);
-  }
-}
