--- src/pubSub/index.js
@@ -1,8 +1,0 @@
-export {
-  configure,
-  subscribe,
-  unsubscribe,
-  publish,
-  getPubSubService,
-} from './interface';
-export { default as PubSubJsService } from './PubSubJsService';
