+++ src/react/AppProvider.jsx
@@ -10,7 +10,7 @@ import { useAppEvent } from './hooks';
 import { identifyAnonymousUser, identifyAuthenticatedUser } from '../analytics';
 import { getAuthenticatedUser, AUTHENTICATED_USER_CHANGED } from '../auth';
 import { getConfig, CONFIG_CHANGED } from '../config';
-import { history } from '../init';
+import { history } from '../initialize';
 import { getLocale, getMessages, IntlProvider, LOCALE_CHANGED } from '../i18n';
 
 const AppProvider = ({ store, children }) => {
